+++ jib-core/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Overloads for `LayerConfiguration#addEntryRecursive` that take providers allowing for setting file permissions/file modification timestamps on a per-file basis ([#1607](https://github.com/GoogleContainerTools/jib/issues/1607))
+
 ### Changed
+
+- `LayerConfiguration` takes file modification time as an `Instant` instead of a `long`
 
 ### Fixed